Subject should be serverspec resource type class
In #325, I wrote "Custom matchers should not check the subject class".

But it's not correct.

Custom matchers should check the class of the argument of match block.

I use ``subject`` as the argument name and it is same as the method name
with RSpec.So it's confusing.

I should not use the name ``subject``, but keep them tentatively.