Do not merge stdout/stderr.
The reason of the merge was the fact that when SSH was requesting a PTY,
stdout/stderr were merged. However, the use of PTY was made
optional (only used for sudo password) in this commit:
 https://github.com/serverspec/specinfra/commit/b4153a7f4f79cfcf5b1d91ebf77da66bd3e76b78

Therefore, we now need to access the true stderr for tests to work as
expected.

The user has to know if he should use stdout or stderr for its
tests. The documentation may state that the use of sudo password will
merge stdout and stderr.